Make sure CPT handles float values

+++ experiments/ConditionalProbabilityTable.cpp
@@ -21,7 +21,7 @@ ConditionalProbabilityTable::ConditionalProbabilityTable(string filePath) {
     const CEREAL_RAPIDJSON_NAMESPACE::Value& array = d["sumOfBrackets"];
     assert(array.IsArray());
     for (unsigned int i = 0; i < VECTOR_SIZE; i++) {
-        this->probabilities.push_back(1.0 * array[i].GetInt() / YEARS);
+        this->probabilities.push_back(1.0 * array[i].GetFloat() / YEARS);
     }
 }
